Rename public screen files to match the component export
