+++ src/index.ts
@@ -1,12 +1,12 @@
 import './util/androidFetch'
 
 export { LoginUiProvider } from './components/publicApi/LoginUiProvider'
-export * from './components/publicApi/PublicChangePasswordScreen'
-export * from './components/publicApi/PublicChangePinScreen'
-export * from './components/publicApi/PublicChangeRecoveryScreen'
-export * from './components/publicApi/PublicLoginScreen'
-export * from './components/publicApi/PublicOtpRepairScreen'
-export * from './components/publicApi/PublicSecurityAlertsScreen'
+export * from './components/publicApi/ChangePasswordScreen'
+export * from './components/publicApi/ChangePinScreen'
+export * from './components/publicApi/ChangeRecoveryScreen'
+export * from './components/publicApi/LoginScreen'
+export * from './components/publicApi/OtpRepairScreen'
+export * from './components/publicApi/SecurityAlertsScreen'
 export * from './components/publicApi/types'
 
 export {
